libtinyrad: updating arguments to tinyrad_initialize()

+++ include/tinyrad.h
@@ -705,7 +705,7 @@ tinyrad_initialize(
          TinyRad **                    trp,
          TinyRadDict *                 dict,
          const char *                  url,
-         uint64_t                      opts );
+         unsigned                      opts );
 
 
 _TINYRAD_F int
